Moved ExtractorVC and its tests inside nl.tudelft.mavensecrets.extractors
