+++ analyzer/src/main/java/nl/tudelft/mavensecrets/extractors/ExtractorVC.java
@@ -1,20 +1,19 @@
-package nl.tudelft;
+package nl.tudelft.mavensecrets.extractors;
 
 import java.util.LinkedList;
 import java.util.List;
 import java.util.Optional;
+
+import nl.tudelft.*;
+import nl.tudelft.Package;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.apache.maven.model.DeploymentRepository;
 import org.apache.maven.model.DistributionManagement;
 import org.apache.maven.model.Model;
 import org.apache.maven.model.Scm;
-import java.io.BufferedReader;
-import java.io.FileReader;
-import java.io.IOException;
-import java.util.jar.JarFile;
 
-public class ExtractorVC implements Extractor{
+public class ExtractorVC implements Extractor {
 
     private static final Logger LOGGER = LogManager.getLogger(ExtractorVC.class);
     Field[] fields;
