+++ analyzer/src/main/resources/config.yml
@@ -4,7 +4,7 @@ extractors:
   - nl.tudelft.mavensecrets.extractors.JavaVersionExtractor
   - nl.tudelft.mavensecrets.extractors.ParentExtractor
   - nl.tudelft.mavensecrets.extractors.PackagingTypeExtractor
-  - nl.tudelft.ExtractorVC
+  - nl.tudelft.mavensecrets.extractors.ExtractorVC
   - nl.tudelft.mavensecrets.extractors.DependencyExtractor
   - nl.tudelft.mavensecrets.extractors.SizeExtractor
   - nl.tudelft.DemoExtractor
